scratch/KrabEar/requirements.txt
mlx-whisper
numpy
sounddevice
requests
scipy
pillow
python-dotenv
pyperclip
pynput
pyobjc-framework-Cocoa; sys_platform == "darwin"
pyobjc-framework-Vision; sys_platform == "darwin"
